Experimental: reuse firstChild and childNodes fields in ContainerNode
In theory this should reduce memory pressure by removing a field
from every object.  In practice it doesn't seem to help very much,
probably because there are already a *bunch* of other fields in every
Element object.  Object sizes might be quantized as well, so we
might not see any improvement until we slimmed down through a
power-of-2 size or something like that.